feat: typescript and registration
