+++ .github/ISSUE_TEMPLATE/bug-report.md
@@ -1,0 +1,47 @@
+---
+name: '🐛 Bug Report'
+about: Did things not work as expected?
+title: "fix: [what is the issue?] in [where is the issue?]"
+labels: 'status:triage'
+---
+
+<!---
+Thanks for filing an issue 😄 ! Before you submit, please read the following:
+Search open/closed issues before submitting. Someone may have reported the same issue before.
+-->
+
+# 🐛 Bug Report
+
+<!--- Provide a general summary of the issue here -->
+
+## 💻 Repro or Code Sample
+
+<!-- Please provide steps to reproduce the issue and/or a code repository, gist, code snippet or sample files -->
+
+## 🤔 Expected Behavior
+
+<!--- Tell us what should happen -->
+
+## 😯 Current Behavior
+
+<!--- Tell us what happens instead of the expected behavior -->
+<!--- If you are seeing an error, please include the full error message and stack trace -->
+<!--- If applicable, provide screenshots -->
+
+## 💁 Possible Solution
+
+<!--- Not obligatory, but suggest a fix/reason for the bug -->
+<!--- Please let us know if you'd be willing to contribute the fix; we'd be happy to work with you -->
+
+## 🔦 Context
+
+<!--- How has this issue affected you? What are you trying to accomplish? -->
+<!--- Providing context helps us come up with a solution that is most useful in the real world -->
+
+## 🌍 Your Environment
+
+<!--- Include as many relevant details as possible about the environment you experienced the bug in -->
+
+* OS & Device: [e.g. MacOS, iOS, Windows, Linux] on [iPhone 7, PC]
+* Browser: [e.g. Microsoft Edge, Google Chrome, Apple Safari, Mozilla FireFox]
+* Version: [e.g. 1.8.0]
