+++ .github/ISSUE_TEMPLATE/feature-request.md
@@ -1,0 +1,39 @@
+---
+name: 🙋 Feature Request
+about: Want us to add something?
+title: "feat: add [what] to/in [where]"
+labels: 'status:triage'
+---
+
+<!---
+Thanks for filing an issue 😄 ! Before you submit, please read the following:
+Search open/closed issues before submitting. Someone may have requested the same feature before.
+-->
+
+# 🙋 Feature Request
+
+<!--- Provide a general summary of the feature here -->
+
+## 🤔 Expected Behavior
+
+<!--- Tell us how the feature should work -->
+
+## 😯 Current Behavior
+
+<!--- Explain how the feature would alter/enhance current behavior -->
+
+## 💁 Possible Solution
+
+<!--- Ideas how to implement this feature -->
+<!--- What implementation solution would be ideal for you? -->
+
+## 🔦 Context
+
+<!--- What are you trying to accomplish? -->
+<!--- How has not having this feature affected you? -->
+<!--- What alternatives have you considered? -->
+
+## 💻 Examples
+
+<!-- Examples help us understand the requested feature better -->
+<!-- Attach screenshots or images if they would add detail to your request -->
