+++ .gitignore
@@ -1,0 +1,6 @@
+# Node
+node_modules
+dist
+
+# macOS
+.DS_Store
