+++ CONTRIBUTING.md
@@ -1,0 +1,33 @@
+# Contributing to Fluent Navigation View Component
+
+Thank you for your interest in contributing to Fluent Navigation View Component!\
+Before you start, please take a moment to review the following guidelines.
+
+### Pull Requests
+When you are ready to submit your changes, please create a pull request.
+
+### Setting up the Project
+
+Initialize repo:
+```cli
+git clone https://github.com/sagemodeninja/fluent-navigation-view-component.git
+cd fluent-navigation-view-component
+```
+
+Install the necessary dependencies: 
+```
+npm install
+```
+
+Build the project:
+```
+npm run build
+```
+
+Run a local demo:
+```
+npm start
+```
+
+### Typescript
+We use TypeScript in this project, a statically typed superset of JavaScript. We recommended that you have a basic understanding of TypeScript. If you are new to TypeScript, we encourage you to check out the official [documentation](https://www.typescriptlang.org/docs/) to get started.
