+++ README.md
@@ -1,11 +1,37 @@
-# Fluent Navigation View Web Component
-
-### Introduction
+# Fluent Navigation View Component
 
 Web component implementation of Windows 11's Navigation View.
 
-### Credits
-#### Authors
+## Documentation
 
-**Gary Antier**\
-[contact@garyantier.com](mailto:contact@garyantier.com).
+See the [wiki](https://github.com/sagemodeninja/fluent-navigation-view-component/wiki) for guides on how to use the components.\
+Or, see the [demo site](https://dev.garyantier.com/fluent/navigation) for examples and demos.
+
+<!-- TODO: Add wiki/documentation. -->
+
+## Contributing
+
+See [CONTRIBUTING.md](CONTRIBUTING.md) for more information.
+
+### Set up
+
+Initialize repo:
+```cli
+git clone https://github.com/sagemodeninja/fluent-navigation-view-component.git
+cd fluent-navigation-view-component
+```
+
+Install the necessary dependencies: 
+```
+npm install
+```
+
+Build the project:
+```
+npm run build
+```
+
+Run a local demo:
+```
+npm start
+```
