--- demo/src/css/style.css
@@ -1,3 +1,0 @@
-* {
-    font-family: 'Segoe UI Variable Text';
-}
