+++ package.json
@@ -1,0 +1,35 @@
+{
+  "name": "fluent-navigation-view-component",
+  "version": "1.3.11",
+  "description": "Web component implementation of Windows 11's Navigation View",
+  "main": "dist/index.bundle.js",
+  "scripts": {
+    "build": "webpack",
+    "start": "webpack serve"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/sagemodeninja/fluent-navigation-view-component.git"
+  },
+  "author": "Gary Antier <contact@garyantier.com> (https://garyantier.com)",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/sagemodeninja/fluent-navigation-view-component/issues"
+  },
+  "homepage": "https://github.com/sagemodeninja/fluent-navigation-view-component#readme",
+  "devDependencies": {
+    "@types/animejs": "^3.1.7",
+    "css-loader": "^6.7.3",
+    "html-webpack-plugin": "^5.5.1",
+    "mini-css-extract-plugin": "^2.7.5",
+    "style-loader": "^3.3.2",
+    "ts-loader": "^9.4.2",
+    "typescript": "^5.0.4",
+    "webpack": "^5.82.0",
+    "webpack-cli": "^5.0.2",
+    "webpack-dev-server": "^4.13.3"
+  },
+  "dependencies": {
+    "animejs": "^3.2.1"
+  }
+}
