+++ public/about.html
@@ -3,9 +3,6 @@
     <head>
         <title>Navigation View - Demo</title>
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <link rel="stylesheet" href="src/fonts/segoe-fluent-icons/segoe-fluent-icons.css">
-        <link rel="stylesheet" href="src/fonts/segoe-ui-variable/segoe-ui-variable.css">
-        <link rel="stylesheet" href="src/css/style.css">
     </head>
     <body>
         <fluent-navigation-view id="navigation_view" header="Home" pane-display-mode="left" header-src="tag" selects-on-load>
@@ -26,7 +23,5 @@
         </fluent-navigation-view>
     </body>
 
-    <script src="src/js/script.js"></script>
     <script src="https://cdn.jsdelivr.net/gh/sagemodeninja/fluent-icon-element-component@v1.0.2/src/fluent-icon-element.min.js"></script>
-    <script src="../src/fluent-navigation-view.js" type="module"></script>
 </html>
