+++ public/contact.html
@@ -3,9 +3,6 @@
     <head>
         <title>Navigation View - Demo</title>
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <link rel="stylesheet" href="src/fonts/segoe-fluent-icons/segoe-fluent-icons.css">
-        <link rel="stylesheet" href="src/fonts/segoe-ui-variable/segoe-ui-variable.css">
-        <link rel="stylesheet" href="src/css/style.css">
     </head>
     <body>
         <fluent-navigation-view id="navigation_view" header="Home" pane-display-mode="left" header-src="tag" selects-on-load>
