+++ public/css/style.css
@@ -1,0 +1,3 @@
+* {
+    font-family: 'Segoe UI Variable';
+}
