+++ public/fonts/segoe-fluent-icons/segoe-fluent-icons.ttf
