+++ public/fonts/segoe-ui-variable/segoe-ui-variable.css
@@ -1,0 +1,6 @@
+@font-face {
+    font-display: fallback;
+    font-family: 'Segoe UI Variable';
+    src: url('segoe-ui-variable.ttf') format('truetype-variations');
+    font-style: normal;
+}
