+++ tsconfig.json
@@ -1,0 +1,9 @@
+{
+	"compilerOptions": {
+		"target": "ES2017",
+		"module": "ES2015",
+		"outDir": "dist",
+		"moduleResolution": "node"
+	},
+	"include": ["src/**/*"]
+}
